Flashlight by Joe Initial Commit
I made this because I wanted to mess around with some Android. I was
also tired of other developers using ads or stealing user’s data. This
code is a little sloppy and untested so updates will surely be coming.
I intend to post an iOS version in the near future as well.